requirements.txt
Flask
gunicorn
pillow
